requirements.txt
BeautifulSoup==3.2.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pyperclip==1.7.0
requests==2.21.0
scipy==1.2.0
selenium==3.141.0
urllib3==1.24.1
v4l2==0.2
